refactored the indentation-sensitive part of the parser into multiple files
